Moved gui commands
