+++ src/main/java/com/dnamaster10/traincartsticketshop/brigadiertest/commands/gui/RemoveGuiEditorCommand.java
@@ -1,24 +1,22 @@
-package com.dnamaster10.traincartsticketshop.brigadiertest.commands;
+package com.dnamaster10.traincartsticketshop.brigadiertest.commands.gui;
 
 import com.dnamaster10.traincartsticketshop.brigadiertest.argumenttypes.GuiNameArgumentType;
 import com.dnamaster10.traincartsticketshop.brigadiertest.argumenttypes.PlayerNameArgumentType;
+import com.dnamaster10.traincartsticketshop.brigadiertest.commands.TicketShopCommand;
 import com.dnamaster10.traincartsticketshop.brigadiertest.suggestions.GuiNameSuggestionProvider;
 import com.dnamaster10.traincartsticketshop.brigadiertest.suggestions.PlayerNameSuggestionProvider;
-import com.dnamaster10.traincartsticketshop.objects.guis.ConfirmGuiTransferGui;
 import com.dnamaster10.traincartsticketshop.util.Players;
 import com.dnamaster10.traincartsticketshop.util.database.accessors.GuiDataAccessor;
+import com.dnamaster10.traincartsticketshop.util.database.accessors.GuiEditorsDataAccessor;
 import com.dnamaster10.traincartsticketshop.util.database.databaseobjects.GuiDatabaseObject;
 import com.dnamaster10.traincartsticketshop.util.database.databaseobjects.PlayerDatabaseObject;
 import com.dnamaster10.traincartsticketshop.util.exceptions.ModificationException;
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.context.CommandContext;
-import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.brigadier.tree.LiteralCommandNode;
 import io.papermc.paper.command.brigadier.CommandSourceStack;
 import io.papermc.paper.command.brigadier.Commands;
-import io.papermc.paper.command.brigadier.argument.ArgumentTypes;
-import io.papermc.paper.command.brigadier.argument.resolvers.selector.PlayerSelectorArgumentResolver;
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.minimessage.MiniMessage;
 import org.bukkit.Bukkit;
@@ -26,27 +24,26 @@ import org.bukkit.entity.Player;
 
 import static com.dnamaster10.traincartsticketshop.TraincartsTicketShop.getPlugin;
 
-public class TransferGuiCommand implements TicketShopCommand {
+public class RemoveGuiEditorCommand implements TicketShopCommand {
     @Override
     public LiteralCommandNode<CommandSourceStack> getRootNode() {
-        return Commands.literal("transfer")
+        return Commands.literal("removeEditor")
                 .requires(ctx -> {
                     if (ctx.getExecutor() instanceof Player player) {
-                        return player.hasPermission("traincartsticketshop.admin.gui.transfer") ||
-                                player.hasPermission("traincartsticketshop.gui.transfer");
+                        return player.hasPermission("traincartsticketshop.gui.removeeditor") || player.hasPermission("traincartsticketshop.admin.gui.removeeditor");
                     }
                     return true;
-                }).then(Commands.argument("id", new GuiNameArgumentType()).suggests(GuiNameSuggestionProvider::getTransferSuggestions)
-                        .then(Commands.argument("player", new PlayerNameArgumentType()).suggests(PlayerNameSuggestionProvider::filterAllNameSuggestions)
+                }).then(Commands.argument("id", new GuiNameArgumentType()).suggests(GuiNameSuggestionProvider::getRemoveEditorSuggestions)
+                        .then(Commands.argument("player", new PlayerNameArgumentType()).suggests(PlayerNameSuggestionProvider::filterGuiEditorSuggestions)
                                 .executes(this::execute))).build();
     }
 
     @Override
     public int execute(CommandContext<CommandSourceStack> ctx) {
-        String guiName = StringArgumentType.getString(ctx, "id");
-        String username = StringArgumentType.getString(ctx, "player");
-
         Bukkit.getScheduler().runTaskAsynchronously(getPlugin(), () -> {
+            String guiName = StringArgumentType.getString(ctx, "id");
+            String editorName = StringArgumentType.getString(ctx, "player");
+
             GuiDataAccessor guiAccessor = new GuiDataAccessor();
             if (!guiAccessor.checkGuiByName(guiName)) {
                 Component component = MiniMessage.miniMessage().deserialize("<red>Gui \"" + guiName + "\" does not exist.");
@@ -54,51 +51,47 @@ public class TransferGuiCommand implements TicketShopCommand {
                 return;
             }
 
-            PlayerDatabaseObject player;
-            try {
-                player = Players.getPlayerByUsername(username);
-            } catch (ModificationException e) {
-                getPlugin().handleSqlException(e);
-                return;
-            }
-
-            if (player == null) {
-                Component component = MiniMessage.miniMessage().deserialize("<red>Player \"" + username + "\" could not be found.");
-                ctx.getSource().getSender().sendMessage(component);
-                return;
-            }
-
             GuiDatabaseObject gui = guiAccessor.getGuiByName(guiName);
 
-            if (gui.ownerUuid().equalsIgnoreCase(player.uuid())) {
-                Component component = MiniMessage.miniMessage().deserialize("<red>Player \"" + player.username() + "\" already owns that Gui.");
-                ctx.getSource().getSender().sendMessage(component);
-                return;
-            }
-
-            if (ctx.getSource().getExecutor() instanceof Player p) {
-                if (!p.hasPermission("traincartsticketshop.admin.gui.transfer")) {
-                    if (!p.getUniqueId().toString().equalsIgnoreCase(gui.ownerUuid())) {
+            if (ctx.getSource().getExecutor() instanceof Player player) {
+                if (!player.hasPermission("traincartsticketshop.admin.gui.removeeditor")) {
+                    if (!player.getUniqueId().toString().equalsIgnoreCase(gui.ownerUuid())) {
                         Component component = MiniMessage.miniMessage().deserialize("<red>You do not own that Gui.");
-                        p.sendMessage(component);
+                        player.sendMessage(component);
                         return;
                     }
                 }
-                getPlugin().getGuiManager().openNewSession(p);
-                ConfirmGuiTransferGui newGui = new ConfirmGuiTransferGui(p, gui.id(), player.uuid());
-                newGui.open();
-                return;
             }
 
+            GuiEditorsDataAccessor editorsAccessor = new GuiEditorsDataAccessor();
+            PlayerDatabaseObject editor;
             try {
-                guiAccessor.updateGuiOwner(gui.id(), player.uuid());
+                editor = Players.getPlayerByUsername(editorName);
             } catch (ModificationException e) {
                 getPlugin().handleSqlException(e);
                 return;
             }
 
-            Component component = MiniMessage.miniMessage().deserialize("<green>Gui \"" + guiName + "\" was transferred to " + player.username() + ".");
-            ctx.getSource().getSender().sendMessage(component);
+            if (editor == null) {
+                Component component = MiniMessage.miniMessage().deserialize("<red>Player \"" + editorName + "\" does not exist");
+                ctx.getSource().getSender().sendMessage(component);
+                return;
+            }
+
+            if (!editorsAccessor.checkGuiEditorByUuid(gui.id(), editor.uuid())) {
+                Component component = MiniMessage.miniMessage().deserialize("<red>Player \"" + editor.username() + "\" is not an editor of that gui.");
+                ctx.getSource().getSender().sendMessage(component);
+                return;
+            }
+
+            try {
+                editorsAccessor.removeGuiEditor(gui.id(), editor.uuid());
+            } catch (ModificationException e) {
+                getPlugin().handleSqlException(e);
+                return;
+            }
+
+            ctx.getSource().getSender().sendMessage("<green>Player \"" + editor.username() + "\" was removed as an editor of Gui \"" + gui.name() + "\"");
         });
         return Command.SINGLE_SUCCESS;
     }
